Fix and Enhance Exploration and EDSM UCs
Removed ImportSphere
Make new UC control FindSystemsUserControl for use by EDSM and
Exploration windows- much more capable device
Fix sorting issues in Bookmarks/combat
panel/journalgrid/ledger/marketdata/missions/commodities/stardistance
EDSM control rewrote to use new UC
Exploration control uses new UC for its import
Exploration control cleaned out lots of crap, recoded, revalidated
EDSM Class more fail safe over getting systems
Numberbox supports a blank look..
Split out Configurable Form and it now supports any control, such as a
UC externally made.  Has a call back to call just before Show.
Themestandard now has some more font helpers
