+++ EDDiscovery/UserControls/UserControlCombatPanel.cs
@@ -224,7 +224,7 @@ namespace EDDiscovery.UserControls
         private void dataGridViewCombat_SortCompare(object sender, DataGridViewSortCompareEventArgs e)
         {
             if (e.Column.Index == 3)        
-                e.SortDataGridViewColumnDate();
+                e.SortDataGridViewColumnNumeric();
         }
