+++ EDDiscovery/UserControls/UserControlEDSM.cs
@@ -33,8 +33,6 @@ namespace EDDiscovery.UserControls
     public partial class UserControlEDSM : UserControlCommonBase
     {
         private string DbColumnSave { get { return "UCEDSM" + ((displaynumber > 0) ? displaynumber.ToString() : "") + "DGVCol"; } }
-        private string DbStar { get { return "UCEDSM" + ((displaynumber > 0) ? displaynumber.ToString() : "") + "Star"; } }
-        private string DbRadius { get { return "UCEDSM" + ((displaynumber > 0) ? displaynumber.ToString() : "") + "Radius"; } }
 
         #region Init
 
@@ -42,7 +40,6 @@ namespace EDDiscovery.UserControls
         {
             InitializeComponent();
             var corner = dataGridViewEDSM.TopLeftHeaderCell; // work around #1487
-            buttonExtEDSMSphere.Enabled = buttonExtDBLookup.Enabled = false;
         }
 
         public override void Init()
@@ -52,13 +49,9 @@ namespace EDDiscovery.UserControls
             dataGridViewEDSM.RowTemplate.Height = 26;
             dataGridViewEDSM.AutoSizeRowsMode = DataGridViewAutoSizeRowsMode.DisplayedCells;     // NEW! appears to work https://msdn.microsoft.com/en-us/library/74b2wakt(v=vs.110).aspx
 
-            textBoxRadius.Text = SQLiteConnectionUser.GetSettingString(DbRadius, "20");
-            textBoxSystemName.Text = SQLiteConnectionUser.GetSettingString(DbStar, "");
-
-            this.textBoxRadius.TextChanged += new System.EventHandler(this.textBoxFilter_TextChanged);
-            this.textBoxSystemName.TextChanged += new System.EventHandler(this.textBoxFilter_TextChanged);
-
-            ValidateEnable();
+            findSystemsUserControl1.Init(displaynumber, true,discoveryform);
+            findSystemsUserControl1.Excel += RunExcel;
+            findSystemsUserControl1.ReturnSystems += StarsFound;
         }
 
         public override void LoadLayout()
@@ -73,163 +66,41 @@ namespace EDDiscovery.UserControls
         public override void Closing()
         {
             DGVSaveColumnLayout(dataGridViewEDSM, DbColumnSave);
-            SQLiteConnectionUser.PutSettingString(DbRadius, textBoxRadius.Text.InvariantParseDoubleNull()!=null ? textBoxRadius.Text : "20");
-            SQLiteConnectionUser.PutSettingString(DbStar, textBoxSystemName.Text);
+            findSystemsUserControl1.Closing();
         }
 
         #endregion
 
-        private void textBoxFilter_TextChanged(object sender, EventArgs e)
+
+        private void StarsFound(List<Tuple<ISystem, double>> systems)       // systems may be null
         {
-            ValidateEnable();
-        }
+            System.Diagnostics.Debug.Assert(Application.MessageLoop);
 
-        void ValidateEnable()
-        { 
-            buttonExtEDSMSphere.Enabled = (textBoxSystemName.Text.Length > 0 && textBoxRadius.Text.InvariantParseDoubleNull() != null);
-            buttonExtDBLookup.Enabled = textBoxSystemName.Text.Length > 0;
-        }
+            dataGridViewEDSM.Rows.Clear();
 
-        #region Excel
-
-        private void buttonExtExcel_Click(object sender, EventArgs e)
-        {
-            if ( dataGridViewEDSM.Rows.Count == 0)
+            if (systems != null)
             {
-                ExtendedControls.MessageBoxTheme.Show(FindForm(), "No data to export", "Export EDSM", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                return;
-            }
+                ISystem cursystem = discoveryform.history.CurrentSystem;        // could be null
+                bool centresort = false;
 
-            Forms.ExportForm frm = new Forms.ExportForm();
-            frm.Init(new string[] { "Export Current View" }, disablestartendtime:true);
-
-            if (frm.ShowDialog(FindForm()) == DialogResult.OK)
-            {
-                if (frm.SelectedIndex == 0)
+                foreach (Tuple<ISystem, double> ret in systems)
                 {
-                    BaseUtils.CSVWriteGrid grd = new BaseUtils.CSVWriteGrid();
-                    grd.SetCSVDelimiter(frm.Comma);
+                    ISystem sys = ret.Item1;
+                    string sep = System.Globalization.CultureInfo.CurrentCulture.NumberFormat.NumberGroupSeparator + " ";
+                    object[] rowobj = {     sys.Name,
+                                            (ret.Item2>=0 ? ret.Item2.ToStringInvariant("0.#") : ""),
+                                            (cursystem != null ? cursystem.Distance(sys).ToString("0.#") : ""),
+                                            sys.X.ToString("0.#") + sep + sys.Y.ToString("0.#") + sep + sys.Z.ToString("0.#"),
+                                            sys.EDSMID.ToStringInvariant() };
 
-                    grd.GetLineStatus += delegate (int r)
-                    {
-                        if (r < dataGridViewEDSM.Rows.Count)
-                        {
-                            return BaseUtils.CSVWriteGrid.LineStatus.OK;
-                        }
-                        else
-                            return BaseUtils.CSVWriteGrid.LineStatus.EOF;
-                    };
-
-                    string[] colh = new string[] { "Star", "Distance", "X", "Y", "Z", "ID" };
-
-                    grd.GetHeader += delegate (int c)
-                    {
-                        return (c < colh.Length && frm.IncludeHeader) ? colh[c] : null;
-                    };
-
-                    grd.GetLine += delegate (int r)
-                    {
-                        DataGridViewRow rw = dataGridViewEDSM.Rows[r];
-                        ISystem sys = rw.Tag as ISystem;
-                        return new Object[]
-                        {
-                            rw.Cells[0].Value, rw.Cells[1].Value, sys.X.ToString("0.#"), sys.Y.ToString("0.#"), sys.Z.ToString("0.#"), rw.Cells[3].Value
-                        };
-                    };
-
-                    if (grd.WriteCSV(frm.Path))
-                    {
-                        if (frm.AutoOpen)
-                            System.Diagnostics.Process.Start(frm.Path);
-                    }
-                    else
-                        ExtendedControls.MessageBoxTheme.Show(FindForm(), "Failed to write to " + frm.Path, "Export Failed", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                    dataGridViewEDSM.Rows.Add(rowobj);
+                    dataGridViewEDSM.Rows[dataGridViewEDSM.Rows.Count - 1].Tag = sys;
+                    centresort |= ret.Item2 >= 0;
                 }
-            }
-        }
 
-        #endregion
-
-        private void buttonExtEDSMSphere_Click(object sender, EventArgs e)
-        {
-            double radius = textBoxRadius.Text.InvariantParseDouble(20);
-            if ( radius > 100 )
-            {
-                if (ExtendedControls.MessageBoxTheme.Show(FindForm(), "This is a large radius, it make take a long time or not work, are you sure?", "Warning - Large radius", MessageBoxButtons.OKCancel, MessageBoxIcon.Asterisk) == DialogResult.Cancel)
-                    return;
+                dataGridViewEDSM.Sort(centresort ? ColumnCentreDistance : ColumnCurrentDistance, ListSortDirection.Ascending);
             }
 
-            Cursor = Cursors.WaitCursor;
-
-            Task taskEDSM = Task<List<Tuple<ISystem, double>>>.Factory.StartNew(() =>
-            {
-                EDSMClass edsm = new EDSMClass();
-                return edsm.GetSphereSystems(textBoxSystemName.Text, radius);
-
-            }).ContinueWith(task => this.Invoke(new Action(() => { SphereGrid(task); })));
-
-        }
-
-        private void SphereGrid(Task<List<Tuple<ISystem, double>>> task)
-        {
-            System.Diagnostics.Debug.Assert(Application.MessageLoop);
-
-            dataGridViewEDSM.Rows.Clear();
-
-            List<Tuple<ISystem, double>> systems = task.Result;
-
-            foreach (Tuple<ISystem, double> ret in systems)
-            {
-                ISystem sys = ret.Item1;
-                string sep = System.Globalization.CultureInfo.CurrentCulture.NumberFormat.NumberGroupSeparator + " ";
-                object[] rowobj = { sys.Name, ret.Item2.ToStringInvariant("0.#"), sys.X.ToString("0.#") + sep + sys.Y.ToString("0.#") + sep + sys.Z.ToString("0.#"), sys.EDSMID.ToStringInvariant() };
-                dataGridViewEDSM.Rows.Add(rowobj);
-                dataGridViewEDSM.Rows[dataGridViewEDSM.Rows.Count - 1].Tag = sys;
-            }
-
-            Cursor = Cursors.Default;
-
-            dataGridViewEDSM.Sort(ColumnDistance, ListSortDirection.Ascending);
-        }
-
-        private void buttonExtDBLookup_Click(object sender, EventArgs e)
-        {
-            Cursor = Cursors.WaitCursor;
-
-            Task taskEDSM = Task<List<ISystem>>.Factory.StartNew(() =>
-            {
-                return SystemClassDB.GetSystemsByName(textBoxSystemName.Text, uselike: true);
-
-            }).ContinueWith(task => this.Invoke(new Action(() => { DBLookup(task); })));
-        }
-
-        private void DBLookup(Task<List<ISystem>> task)
-        {
-            System.Diagnostics.Debug.Assert(Application.MessageLoop);
-
-            dataGridViewEDSM.Rows.Clear();
-
-            List<ISystem> systems = task.Result;
-
-            HistoryEntry helast = discoveryform.history.GetLast;
-            ISystem home = helast != null ? helast.System : new SystemClass("Sol", 0, 0, 0);
-
-            foreach (ISystem sys in systems)
-            {
-                string sep = System.Globalization.CultureInfo.CurrentCulture.NumberFormat.NumberGroupSeparator + " ";
-
-                string dist = "";
-                if (sys.HasCoordinate)
-                    dist = sys.Distance(home).ToString("0.#");
-
-                object[] rowobj = { sys.Name, dist, sys.X.ToString("0.#") + sep + sys.Y.ToString("0.#") + sep + sys.Z.ToString("0.#"), sys.EDSMID.ToStringInvariant() };
-                dataGridViewEDSM.Rows.Add(rowobj);
-                dataGridViewEDSM.Rows[dataGridViewEDSM.Rows.Count - 1].Tag = sys;
-            }
-
-            Cursor = Cursors.Default;
-
-            dataGridViewEDSM.Sort(ColumnStar, ListSortDirection.Ascending);
         }
 
         ISystem rightclicksystem = null;
@@ -301,9 +172,70 @@ namespace EDDiscovery.UserControls
 
         private void dataGridViewEDSM_SortCompare(object sender, DataGridViewSortCompareEventArgs e)
         {
-            if (e.Column.Index == 1 || e.Column.Index == 3)
-                e.SortDataGridViewColumnDate();
+            if (e.Column.Index == 1 || e.Column.Index == 2 || e.Column.Index == 4)
+                e.SortDataGridViewColumnNumeric();
            
         }
+
+        #region Excel
+
+        private void RunExcel()
+        {
+            if (dataGridViewEDSM.Rows.Count == 0)
+            {
+                ExtendedControls.MessageBoxTheme.Show(FindForm(), "No data to export", "Export EDSM", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                return;
+            }
+
+            Forms.ExportForm frm = new Forms.ExportForm();
+            frm.Init(new string[] { "Export Current View" }, disablestartendtime: true);
+
+            if (frm.ShowDialog(FindForm()) == DialogResult.OK)
+            {
+                if (frm.SelectedIndex == 0)
+                {
+                    BaseUtils.CSVWriteGrid grd = new BaseUtils.CSVWriteGrid();
+                    grd.SetCSVDelimiter(frm.Comma);
+
+                    grd.GetLineStatus += delegate (int r)
+                    {
+                        if (r < dataGridViewEDSM.Rows.Count)
+                        {
+                            return BaseUtils.CSVWriteGrid.LineStatus.OK;
+                        }
+                        else
+                            return BaseUtils.CSVWriteGrid.LineStatus.EOF;
+                    };
+
+                    string[] colh = new string[] { "Star", "Centre Distance", "Current Distance", "X", "Y", "Z", "ID" };
+
+                    grd.GetHeader += delegate (int c)
+                    {
+                        return (c < colh.Length && frm.IncludeHeader) ? colh[c] : null;
+                    };
+
+                    grd.GetLine += delegate (int r)
+                    {
+                        DataGridViewRow rw = dataGridViewEDSM.Rows[r];
+                        ISystem sys = rw.Tag as ISystem;
+                        return new Object[]
+                        {
+                            rw.Cells[0].Value, rw.Cells[1].Value, rw.Cells[2].Value, sys.X.ToString("0.#"), sys.Y.ToString("0.#"), sys.Z.ToString("0.#"), rw.Cells[4].Value
+                        };
+                    };
+
+                    if (grd.WriteCSV(frm.Path))
+                    {
+                        if (frm.AutoOpen)
+                            System.Diagnostics.Process.Start(frm.Path);
+                    }
+                    else
+                        ExtendedControls.MessageBoxTheme.Show(FindForm(), "Failed to write to " + frm.Path, "Export Failed", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                }
+            }
+        }
+
+        #endregion
+
     }
 }
