+++ ExtendedControls/ExtendedControls.csproj
@@ -93,6 +93,9 @@
     <Compile Include="Controls\TreeViewCustom.Designer.cs">
       <DependentUpon>TreeViewCustom.cs</DependentUpon>
     </Compile>
+    <Compile Include="Forms\ConfigurableForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Forms\SmartSysMenuForm.cs">
       <SubType>Form</SubType>
     </Compile>
